@@ -6,7 +6,6 @@
     "dev-bot": "./dist/src/cli/cli.js"
   },
   "scripts": {
-    "postinstall": "typings install",
     "build": "rimraf dist/* && tsc",
     "test": "npm run build && mocha 'dist/test/**/*-test.js'",
     "prepublish": "npm run build"
@@ -33,19 +32,30 @@
   },
   "homepage": "https://github.com/pimterry/dev-bot-tool#readme",
   "devDependencies": {
+    "@types/chai": "^3.4.34",
+    "@types/mocha": "^2.2.32",
+    "@types/sinon": "^1.16.31",
+    "@types/sinon-chai": "^2.7.27",
+    "@types/temp": "^0.8.29",
     "chai": "^3.5.0",
     "mocha": "^3.0.2",
     "rimraf": "^2.5.4",
     "sinon": "^1.17.5",
     "sinon-chai": "^2.8.0",
     "temp": "^0.8.3",
-    "typescript": "^1.8.10",
-    "typings": "^1.3.3"
+    "typescript": "^2.1.0-dev.20161019"
   },
   "engines": {
     "node": ">=4.0.0"
   },
   "dependencies": {
+    "@types/aws-lambda": "0.0.3",
+    "@types/aws-sdk": "0.0.38",
+    "@types/docopt": "^0.6.22-alpha",
+    "@types/dotenv": "^2.0.18",
+    "@types/fs-extra": "0.0.33",
+    "@types/jszip": "0.0.31",
+    "@types/node": "^6.0.45",
     "aws-sdk": "^2.6.0",
     "dev-bot": "^0.2.0",
     "docopt": "^0.6.2",
